requirements.txt
python-telegram-bot>=22.0
pandas>=2.2.0
numpy>=2.2.0